reading data does not work when creating an object in the database

+++ src/main/java/com/Car_Rental_Spring/controller/requests/worker/WorkerUserCreateRequest.java
@@ -15,7 +15,7 @@ import javax.validation.constraints.Size;
 public class WorkerUserCreateRequest {
 
     @Size(min = 1)
-    private Long idUser;
+    private int idUser;
 
     @Size (min = 1)
     private String nameWork;
