+++ src/main/java/com/Car_Rental_Spring/repository/RoleDao.java
@@ -4,8 +4,8 @@ import com.Car_Rental_Spring.entity.MRoles;
 
 import java.util.List;
 
-public interface RoleDao extends GenericDao<MRoles, Long> {
+public interface RoleDao extends GenericDao<MRoles, Long > {
     MRoles findRole(String str);
-    List<MRoles> getRolesByUserId(Long userId);
+    List<MRoles> getRolesByUserId(int userId);
 
 }
